+++ portfolio/src/main/webapp/index.html
@@ -29,7 +29,7 @@
       <div class="comments">
         <h2>Comments</h2>
         <div id="login-message"></div>
-        <form id="comments-form" action="/comments" method="POST">
+        <form id="comments-form">
           <input type=hidden name="email-input" id="email-input"/>
           <p>Enter your comment here:</p>
           <textarea name="comment-input" id="comment-input"></textarea>
@@ -45,7 +45,7 @@
             <option value="👽">👽</option>
           </select>
           <br></br>
-          <input type="submit"/>
+          <input type="button" value="Submit" onclick="addComment()"/>
         </form>
         <hr>
         <div id="text-container"></div>
